docs: add hints to customize configuration
By default the starter script will only use the configuration files if
no other configuration is provided, otherwise only the provided
configuration is used. This isn't clear enough in the doc, leading to
misconfigured environments. This update adds a few more details about
how the configuration is built and how it can be safely customized.